WIP on fixing samples.
Pass assert to test cases.
