+++ package.json
@@ -26,10 +26,12 @@
   "devDependencies": {
     "babel-core": "^5.8.20",
     "babelify": "^6.1.3",
+    "browserify-versionify": "^1.0.6",
     "chai": "^3.2.0",
     "es6-module-loader": "^0.17.3",
     "karma": "^0.13.8",
     "karma-chrome-launcher": "^0.2.0",
+    "markdown": "^0.5.0",
     "sinon": "^1.15.4"
   },
   "main": "src/Up.js",
