+++ sam/Fun.js
