+++ sam/Tem.js
