--- sam/count.js
@@ -1,8 +1,0 @@
-import {UpCount, T} from '../src/Up';
-var dom;
-
-//===START
-dom = T(UpCount());
-//===END
-
-export default dom;
