+++ sam/style.js
@@ -1,14 +1,63 @@
-import {E, U, T} from '../src/Up';
-var dom;
+import {U, UpStyle} from '../src/Up';
+var {assign, keys} = Object;
 
-//===START
-var style = U({ backgroundColor: 'pink' });
-function click() { style.backgroundColor = 'cyan'; }
+var theme = U({
+});
 
-dom = E('button') .
-  is({ style }) .
-  does({ click }) .
-  has(T("Change my background"));
-//===END
+var setTheme = t => assign(theme, themes[t]);
+var getThemeNames = _ => keys(themes);
 
-export default dom;
+var themes = {
+  sunset: {
+    bodyBackgroundColor: "wheat"
+  }
+};
+
+setTheme("sunset");
+
+UpStyle([
+
+  ["body", {
+    fontFamily : 'sans-serif',
+    backgroundColor: theme.bodyBackgroundColor
+  }],
+
+  ["h3", {
+    backgroundColor: 'brown',
+    color: 'white',
+    padding: '6px'
+  }],
+
+  [".code", {
+    whiteSpace: 'pre',
+    fontFamily: 'monospace',
+    backgroundColor: 'pink',
+    margin: '12px 40px',
+    padding: '12px',
+    fontSize: 'larger'
+  }],
+
+  [".result", {
+    backgroundColor: 'beige',
+    margin: '12px 40px',
+    padding: '12px'
+  }],
+
+  [".hide", {
+    display: 'none'
+  }],
+
+  ["code", {
+    fontSize: 'larger',
+    backgroundColor: 'lightgray',
+    border: "1px solid gray",
+    paddingLeft: "0.2em",
+    paddingRight: "0.2em"
+  }]
+
+]);
+
+export {
+  setTheme,
+  getThemeNames
+}
