+++ tst/index.html
@@ -17,17 +17,8 @@
     <script src="../node_modules/chai/chai.js"></script>
     <script src="vendor/sinon-1.15.4.js"></script>
 
-    <!-- Babel transpiler -->
-    <!--<script src="../node_modules/babel-core/browser.js"></script>-->
-
-    <!-- Module loader -->
-    <script src="../node_modules/es6-module-loader/dist/es6-module-loader-dev.js"></script>
-
     <!-- Babel polyfill, needed for generators. -->
     <script src="../node_modules/babel-core/browser-polyfill.js"></script>
-
-    <!-- Configure es6-module loader. -->
-    <script src="../es6-module-loader-config.js"></script>
 
     <!-- livereload -->
     <script>
